feat: gemini code review반영
