+++ stores/user-store.ts
@@ -1,4 +1,4 @@
-import { RankingData } from '@/types/auth';
+import { RankingData } from '@/types/user';
 import { create } from 'zustand';
 import * as userAPI from '../api/user.service';
 
@@ -8,6 +8,13 @@ interface UserState {
 
 export const useUserStore = create<UserState>((set) => ({
   userRanking: async (type: string) => {
+    const response = await userAPI.UserRanking(type);
+    if (!response) {
+      return {
+        currentUserRank: null,
+        list: [],
+      };
+    }
     const { currentUserRank, list } = await userAPI.UserRanking(type);
     return {
       currentUserRank,
